backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
qrcode[pil]
passlib==1.7.4
bcrypt==3.2.2
python-jose[cryptography]